feat(routes): add consign for route management
