+++ src/app.js
@@ -1,27 +1,17 @@
 const express = require('express');
+const consign = require('consign');
 
 const port = 3001;
-
 const app = express();
-app.use(express.urlencoded({extended: true})); 
-app.use(express.json());
 
-var users = [
-  {name:'John Doe', email: 'email@email.com' }
-]
+consign({ cwd: 'src', verbose: false})
+  .include('./config/middlewares.js')
+  .then('./routes')
+  .then('./config/routes.js')
+  .into(app)
 
 app.get('/', (req, res) => {
   res.status(200).send()
-});
-
-app.get('/users', (req, res) => {  
-  res.status(200).json(users);
-});
-
-app.post('/users', (req, res) => {
-  console.log('req.body', req.body);
-  users.push(req.body);
-  res.status(201).json(req.body);
 });
 
 module.exports = app ;
