+++ src/config/middlewares.js
@@ -1,0 +1,8 @@
+const express = require('express');
+
+module.exports = (app) => {
+
+app.use(express.urlencoded({extended: true})); 
+app.use(express.json());
+
+}
