+++ src/config/routes.js
@@ -1,0 +1,5 @@
+module.exports = (app) => {
+  app.route('/users')
+  .get(app.routes.users.findAll)
+  .post(app.routes.users.create);
+}
